edit dashboard and profile functionality

+++ app/Http/Controllers/AdminCtrls/BrandsController.php
@@ -87,7 +87,11 @@ class BrandsController extends Controller
     public function edit($id)
     {
         $brand = brands::find($id);
-        return view('brands.edit')->with('brand',$brand);
+        if(Auth::user()->id === $brand->admin_id || Auth::user()->super_admin === 1 ){
+            return view('brands.edit')->with('brand',$brand);
+        }else {
+            return redirect('/admin/brands')->with('error', 'This action for owner or Super admin only');
+        }
     }
 
     /**
@@ -99,19 +103,23 @@ class BrandsController extends Controller
      */
     public function update(Request $request, $id)
     {
-        $this->validate($request, [
-            'brand'=>'required|min:2|unique:brands,brand_name,'.$id,
-            'description'=>'required|min:5',
-        ]);
-
-        $brand = brands::find($id);
-        $brand->brand_name = $request->input('brand');
-        $brand->description= $request->input('description');
-        $brand->admin_id = Auth::user()->id;
-        $brand->save();
-        return redirect('admin/brands')->with('success', 'Item has been updated successfully');
+        if(Auth::user()->id === $brand->admin_id || Auth::user()->super_admin === 1 ){
+            $this->validate($request, [
+                'brand'=>'required|min:2|unique:brands,brand_name,'.$id,
+                'description'=>'required|min:5',
+                ]);
+            
+            $brand = brands::find($id);
+            $brand->brand_name = $request->input('brand');
+            $brand->description= $request->input('description');
+            $brand->admin_id = Auth::user()->id;
+            $brand->save();
+            return redirect('admin/brands')->with('success', 'Item has been updated successfully');
+        }else {
+            return redirect('/admin/brands')->with('error', 'This action for owner or Super admin only');
+        }
     }
-
+    
     /**
      * Remove the specified resource from storage.
      *
@@ -121,25 +129,39 @@ class BrandsController extends Controller
     public function destroy($id)
     {
         $brand = brands::find($id);
-        $brand->delete();
-        return redirect('admin/brands')->with('success', 'Item has been deleted successfully');
+        if(Auth::user()->id === $brand->admin_id || Auth::user()->super_admin === 1 ){
+            $brand->delete();
+            return redirect('admin/brands')->with('success', 'Item has been deleted successfully');
+        }else {
+            return redirect('/admin/brands')->with('error', 'This action for owner or Super admin only');
+        }
     }
 
 
     public function activate(Request $request, $id)
     {
         $brand = brands::find($id);
-        $brand->active = '1';
-        $brand->save();
-        return redirect('admin/brands')->with('success', 'Item has been activated successfully');
+        if(Auth::user()->id === $brand->admin_id || Auth::user()->super_admin === 1 ){
+            $brand->active = '1';
+            $brand->save();
+            return redirect('admin/brands')->with('success', 'Item has been activated successfully');
+        }else {
+            return redirect('/admin/brands')->with('error', 'This action for owner or Super admin only');
+        }
+        
     }
 
     public function inActivate(Request $request, $id)
     {
         $brand = brands::find($id);
-        $brand->active = '0';
-        $brand->save();
-        return redirect('admin/brands')->with('success', 'Item has been inActivated successfully');
+        if(Auth::user()->id === $brand->admin_id || Auth::user()->super_admin === 1 ){
+            $brand->active = '0';
+            $brand->save();
+            return redirect('admin/brands')->with('success', 'Item has been inActivated successfully');
+        }else {
+            return redirect('/admin/brands')->with('error', 'This action for owner or Super admin only');
+        }
+        
     }
 
     public function search(Request $request)
@@ -152,4 +174,12 @@ class BrandsController extends Controller
         return view('brands.search')->with(['brands'=>$brands]);
         
     }
+    // public function forSuperAdmin($route)
+    // {
+    //     if(Auth::user()->id === $brand->admin_id || Auth::user()->super_admin === 1 ){
+    //         return true;
+    //     }else {
+    //         return redirect('/admin/'.$route)->with('error', 'This action for owner or Super admin only');
+    //     }
+    // }
 }
